third_party/WebKit/LayoutTests/fast/dom/custom/registration-context-delete-then-register-expected.txt

Tests destroying a context, then trying to register a new custom element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS destroyed context
PASS caught expected exception "InvalidStateError: Failed to call 'register' on 'Document' for type 'x-a': the context is no longer valid."
PASS child done
PASS successfullyParsed is true

TEST COMPLETE
